Port to use React Proxy (0.5.0-alpha)
