+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-hot-api",
-  "version": "0.4.6",
+  "version": "0.5.0-alpha",
   "description": "A generic library implementing hot reload for React components without unmounting or losing their state.",
   "main": "modules/index.js",
   "scripts": {
@@ -17,9 +17,12 @@
   },
   "homepage": "https://github.com/gaearon/react-hot-api",
   "peerDependencies": {
-    "react": ">=0.11.0 || 0.13.0-beta.1 || 0.13.0-rc1 || 0.13.0-rc2 || ^0.14.0-alpha"
+    "react": ">=0.13.0 || ^0.14.0-alpha"
   },
   "devDependencies": {
     "webpack": "1.4.8"
+  },
+  "dependencies": {
+    "react-proxy": "^0.6.1"
   }
 }
